requirements.txt
flask==3.0.0
psycopg2-binary==2.9.11
redis==5.0.1
neo4j==5.20.0
python-dotenv==1.0.0
